Fix HUD order display
Trying to keep track of the current index is a mess. Dropping this
requirement allows for a linked list to be used, which makes it very
easy to get the next order in the sequence.

Assign and display a unique ID for each order to make up for the removal
of the current order index.

Remove the total order count temporarily. This will return when the UI
is redesigned to accomodate for the order timers.

Also refactor the way orders are stored. Instead of a class which pairs
the ScriptableObject and the time, add the time as a field of the SO.
Create a clone of the SO for each order added.